topCharts/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
flask ~= 3.0.2
pytest ~= 5.4
mock
prometheus_client
time